% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclean.R
\name{dataclean}
\alias{dataclean}
\title{dataclean}
\usage{
dataclean(X.c, X.o, y)
}
\arguments{
\item{X.c}{data frame that has count data from all levels (only count data will be row scaled)}

\item{X.o}{data frame that has other environmental variables (no scaling will be done, those variables will scaled together with proportion data in LASSO step)}

\item{y}{a vector representing the outcome (0 or 1 for binomial model)}
}
\value{
data.Xy: a dataframe containing all variables named as X1,X2,...,Xp and the binary outcome (called status)in the last column; this data frame will be used in the other functions for data analysis
}
\description{
This function is to 1)Scale the count data (count data only) to proportion 2)create a data frame consisting of proportion data, and 3) Keep an variable name list (original variable names and names in terms of X's, e.g.X1,X2,..,etc. )
#environmental data (host genome and other information about observations)
}
