% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{Ranking}
\alias{Ranking}
\title{Ranking}
\usage{
Ranking(data, model)
}
\arguments{
\item{data}{the data object return from the dataclean function (the last column is the outcome)}

\item{model}{model object from sub-sampling procedure}
}
\value{
table: a table shows the ranked variable list  with its frequency (descending order)

Beta: coefficients flag (1 or 0) indicating if the variable is selected; intercept is not included;
}
\description{
This function is to rank the variables after B subsampings;It also removes the highly correlated variables from lower level
}
