% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectnew.R
\name{selectnew}
\alias{selectnew}
\title{selectnew}
\usage{
selectnew(
  vslist,
  ranktable,
  data,
  weights,
  ncores = 1,
  family,
  alpha_l,
  alpha_u,
  C
)
}
\arguments{
\item{ranktable}{:ranking table from ranking step}

\item{family}{:generalized model families}

\item{alpha_l}{is the minimum significance level(>=0)}

\item{data}{the dataframe should be arranged in the way such that columns are X1,X2,X3....,Xp, status.  Where Xi's are variables and status is the outcome(for the logistic regression, the outcome is in terms of 0/1)}

\item{ncores}{no of parallel computing cores}

\item{C}{the number of permutation times}

\item{alpha_u}{the upper significance level}

\item{weights}{In a binomial model, weights: =TRUE: if weighted version is desired; =FALSE, otherwise ; In other models,weights: =vector of weights of the same size as the sample size N: if weighted version is desired;=FALSE, otherwise (other generalized model)}

\item{vslist}{ the current list of selected variables}

}
\value{
vslist: the updated list of selected variables

dev.dist: deviance distributions used for selecting the new variable

vtlist has 1)alpha.range for the newly selected variable,2)selvar: the newly selected variable name,3)pval: pvalue for the newly selected varaible,and 4)dev: the deviance value contributed by the newly selected variable
}
\description{
This function is to add new node (new deviance distribution for adding the new variable)
}
