% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selpath.R
\name{selpath}
\alias{selpath}
\title{selpath}
\usage{
selpath(data, weights, ranktable, ncores, family, C, alpha_u)
}
\arguments{
\item{ranktable}{:ranking table from ranking step}

\item{family}{:generalized model families}

\item{data}{the dataframe should be arranged in the way such that columns are X1,X2,X3....,Xp, status.  Where Xi's are variables and status is the outcome(for the logistic regression, the outcome is 0/1)}

\item{ncores}{no of parallel computing cores}

\item{C}{the number of permutation times}

\item{alpha_u}{the upper significance level}

\item{weights}{In a binomial model, weights: =TRUE: if weighted version is desired; =FALSE, otherwise ; In other models,weights: =vector of weights of the same size as the sample size N: if weighted version is desired;=FALSE, otherwise (other generalized model)}
}
\value{
selpoint: a list. it contains each selected variable point,information includes 1)vslist: the variable sect before selecting this variable listed in 'selvar' 2)alpha.range: the variable will be selected within this alpha range 3)pval: pvalue of the variable 4)selvar:selected variable 5)vslist:variable sect after selecting the variable listed in 'selvar'

sel.nodes: a list. deviance distributions used for selecting the new variable; it includes 1)vslist: the variable sect before the new selection 2)dev.dist: the permutation for selecting the new variable 3)vtlist that has  i)pval: pvalue of the proposed variable ii)selvar: selected variable (the proposed variable is NULL if not selected) iii)dev:deviance contributed by the proposed variable
}
\description{
#This function is to trace the selection path
}
