% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selvar_alpha.R
\name{selvar_alpha}
\alias{selvar_alpha}
\title{selvar_alpha}
\usage{
selvar_alpha(res, alpha)
}
\arguments{
\item{res}{'mod' object returned from 'selpath' function}

\item{alpha}{alpha level(default alpha=0.05)(a single value up to the value 'alpha_u' sepecified in selpath() function)}
}
\value{
selvar:final selected variable

pval:pvalue of each selected variable (present if at least 1 var is selected)

devlist:deviance contributed by each selected variable (given the previous selected variables;present if at least 1 var is selected )

dist.mat:a list of deviance permutation distributions (including the distribution from the step from which no more variable is added)
}
\description{
This function is to extract summarize the results at 'alpha' level from 'mod' object to obtain 1)final selected variables 'selvar' #2)pvalue of each selected variable according to the variables in the 'selvar' 3)deviance contributed by each selected variable (given the previous selected variables 4)deviance permutation distribution 5)cutoff value based on (1-alpha)percentile of the deviance permutation distribution;when no variable is selected, only return the last deviance distribution and the cutoff value;this function can be used separately after running selpath(); the alpha value must be >0 and <= alpha_u parameter from SURF()
}
