% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.dev_cox.R
\name{update_dev_cox}
\alias{update_dev_cox}
\title{update_dev_cox}
\usage{
update_dev_cox(data, vslist, C, ncores, weights)
}
\arguments{
\item{data}{the variable 'data' within seqcutoff()}

\item{vslist}{a vector of selected variables}

\item{C}{the number of permutation times}

\item{weights}{=TRUE: if weighted version is desired,   =FALSE, otherwise (binomial model); weights: =vector of weights of the same size as the sample size N: if weighted version is desired,   =FALSE, otherwise (other generalized model)}

\item{ncores}{the number of cores to use for parallel computation}

}
\value{
dev: a vector of deviance after C permutations (length OF this vector is C)
}
\description{
For COX proportional model ONLY. This function is to derive the deviance distribution based on the permutation method;This function is not to be used independently but will be called by the function selectnew()
}
